Modified moved files so they communicate properly
