+++ Core/monitorManager.py
@@ -1,4 +1,5 @@
 import logger
+
 logging = logger.Logger(__file__)
