+++ GUI/activeMonitorsWidget.py
@@ -1,7 +1,7 @@
 from PyQt5 import QtCore, QtGui, QtWidgets
 from PyQt5.QtCore import Qt
 import time
-from Core import logger
+import logger
 
 logging = logger.Logger(__file__)
